+++ src/main/java/de/theidler/create_mobile_packages/toast/CustomToast.java
@@ -11,11 +11,17 @@ public class CustomToast {
     public final Component subtitle;
     public final ItemStack icon;
     public final long lastUpdate = System.currentTimeMillis();
+    public final int timeout;
 
-    public CustomToast(UUID uuid, Component title, Component subtitle, ItemStack icon) {
+    public CustomToast(UUID uuid, Component title, Component subtitle, ItemStack icon, int timeout) {
         this.uuid = uuid;
         this.title = title;
         this.subtitle = subtitle;
         this.icon = icon;
+        this.timeout = timeout;
+    }
+
+    public CustomToast(UUID uuid, Component title, Component subtitle, ItemStack icon) {
+        this(uuid, title, subtitle, icon, 5000);
     }
 }